Action: Renamed all of the pages, implemented their own scss files, broke out Card component for home page, named a few section components
