+++ frontend/src/components/atoms/footer/Footer.jsx
@@ -1,8 +1,9 @@
 import React from "react";
+import "./Footer.scss"
 
 const Footer = () => {
   return (
-    <footer className="footer bg-dark">
+    <footer className="dkFooter bg-dark">
       <div className="container text-center">
         <span className="text-muted">Landon McKell | Web Developer</span>
       </div>
