+++ frontend/src/components/atoms/footer/Footer.scss
@@ -1,0 +1,6 @@
+.dkFooter {
+  width: 100%;
+  height: 60px;
+  line-height: 60px;
+  background-color: #f5f5f5;
+}
