+++ frontend/src/components/sectionTemplates/techStackSection/TechStackSection.scss
@@ -1,0 +1,6 @@
+.dkTechStackSection {
+  .skills-section .card hr {
+    border-width: 0.25rem;
+    width: 3rem;
+  }
+}
