+++ frontend/src/pages/projectDetailsPage/ProjectDetailsPage.scss
@@ -1,0 +1,5 @@
+.dkProjectDetailsPage {
+    .details-btn {
+        min-width: 7em;
+      }
+}
